🚧  R03 : Typescript translation progress (utils files)
