--- src/utils/ref.js
@@ -1,4 +1,0 @@
-export function focusTitleEditBox(ref) {
-  if (!ref.current) return;
-  ref.current.focus();
-}
