+++ src/utils/ref.tsx
@@ -1,0 +1,4 @@
+export function focusTitleEditBox(ref: React.RefObject<HTMLInputElement>) {
+  if (!ref.current) return;
+  ref.current.focus();
+}
