--- src/utils/window.js
@@ -1,15 +1,0 @@
-export const refreshPage = () => {
-  window.location.reload();
-};
-
-export const storeLocally = (username, key, value) => {
-  if (!username) return;
-  localStorage.setItem(username, JSON.stringify({ [key]: value }));
-};
-
-export const retrieveFromLocalStorage = (username, key) => {
-  if (!username) return;
-  const storedData = JSON.parse(localStorage.getItem(username));
-  if (!storedData) return;
-  return storedData[key];
-};
